# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TraclyTag Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- PRODUCTS ----------
  /products:
    get:
      operationId: listProducts
      tags: [products]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- LOCATIONS ----------
...
  /reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockRow"
  /reports/product:
    get:
      operationId: getProductReport
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductReportRow"
  /reports/shipper-summary:
    get:
      operationId: getShipperSummary
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipperSummaryRow"
  /reports/pallet-summary:
    get:
      operationId: getPalletSummary
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PalletSummaryRow"
...
components:
  schemas:
...
    Role:
      type: string
      enum: [master, client_admin, operator]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/Role"
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, phone, role, companyId, companyName, createdAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        password:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        companyId:
          type: ["integer", "null"]
      required: [username, email, password, role]
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        skuId:
          type: string
        name:
          type: string
        skuSize:
          type: string
        marketedBy:
          type: string
        sapDescription:
          type: ["string", "null"]
        gtin:
          type: string
        mrp:
          type: number
        registrationNo:
          type: ["string", "null"]
        l1Size:
          type: integer
          description: Units per 1st level inner pack
        l2Size:
          type: integer
          description: 1st level packs per 2nd level inner pack
        shipperSize:
          type: integer
          description: 2nd level packs per shipper
        cautionLogoUrl:
          type: ["string", "null"]
        productLogoUrl:
          type: ["string", "null"]
        labelPdfUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - skuId
        - name
        - skuSize
        - marketedBy
        - sapDescription
        - gtin
        - mrp
        - registrationNo
        - l1Size
        - l2Size
        - shipperSize
        - cautionLogoUrl
        - productLogoUrl
        - labelPdfUrl
        - createdAt
    CreateProductBody:
      type: object
      properties:
        skuId:
          type: string
        name:
          type: string
        skuSize:
          type: string
        marketedBy:
          type: string
        sapDescription:
          type: ["string", "null"]
        gtin:
          type: string
          description: 14 digit GTIN
        mrp:
          type: number
        registrationNo:
          type: ["string", "null"]
        l1Size:
          type: integer
        l2Size:
          type: integer
        shipperSize:
          type: integer
        cautionLogoUrl:
          type: ["string", "null"]
        productLogoUrl:
          type: ["string", "null"]
        labelPdfUrl:
          type: ["string", "null"]
      required:
        - skuId
        - name
        - skuSize
        - marketedBy
        - gtin
        - mrp
        - l1Size
        - l2Size
        - shipperSize
...
    StockRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        totalCodes:
          type: integer
        mapped:
          type: integer
        unmapped:
          type: integer
        location:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - totalCodes
        - mapped
        - unmapped
        - location
    ProductReportRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        mapped:
          type: integer
        unmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - mapped
        - unmapped
    ShipperSummaryRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        shipperMapped:
          type: integer
        shipperUnmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - shipperMapped
        - shipperUnmapped
    PalletSummaryRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        palletMapped:
          type: integer
        palletUnmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - palletMapped
        - palletUnmapped
...