# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  phone: string | null;
  role: Role;
  /** @nullable */
  companyId: number | null;
  /** @nullable */
  companyName: string | null;
  createdAt: Date;
}